requirements.txt
distro==1.9.0 
fastapi==0.109.2 
httpcore==1.0.3 
httpx==0.26.0 
openai==1.12.0 
pydantic==2.6.1 
pydantic-core==2.16.2 
python-dotenv==1.0.1 
starlette==0.36.3 
tqdm==4.66.2 
uvicorn==0.27.1
requests==2.31.0